chore(extension): update dependencies and refactor environment utilities
- Add Antd and i18next-browser-languagedetector to extension dependencies
- Refactor environment-related utilities across packages
- Update WXT configuration to optimize build and dependencies
- Consolidate runtime and environment detection logic in utils package
- Improve cross-package imports and type handling
