+++ apps/extension/package.json
@@ -73,7 +73,9 @@
     "vite-plugin-encoding": "~1.0.0",
     "vite-tsconfig-paths": "~4.3.2",
     "xpath-to-selector": "^1.1.3",
-    "zustand": "~4.5.1"
+    "zustand": "~4.5.1",
+    "antd": "^5.21.5",
+    "i18next-browser-languagedetector": "^8.0.2"
   },
   "devDependencies": {
     "@arco-plugins/vite-react": "^1.3.3",
@@ -88,6 +90,6 @@
     "sass": "1.71.1",
     "tailwind-merge": "^2.2.2",
     "tailwindcss": "^3.2.6",
-    "wxt": "^0.19.26"
+    "wxt": "workspace:*"
   }
 }
