+++ packages/ai-workspace-common/src/utils/env.ts
@@ -1,87 +1,4 @@
-import { IRuntime } from '@refly/common-types';
-
-export enum IENV {
-  PRODUCTION = 'production',
-  STAGING = 'staging',
-  TEST = 'test',
-  DEVELOPMENT = 'development',
-}
-
-export const getEnv = () => {
-  const env = process.env.NODE_ENV;
-
-  return env;
-};
-
-declare global {
-  interface Window {
-    ENV?: {
-      API_URL: string;
-      COLLAB_URL: string;
-    };
-  }
-}
-
-export const serverOrigin = import.meta.env.VITE_API_URL || window.ENV?.API_URL;
-export const wsServerOrigin = import.meta.env.VITE_COLLAB_URL || window.ENV?.COLLAB_URL;
-
-let runtime: IRuntime;
-
-export const getRuntime = () => {
-  return runtime;
-};
-
-export const setRuntime = (r: IRuntime) => {
-  runtime = r;
-};
-
-export class ReflyEnv {
-  getExtensionVersion() {
-    return chrome.runtime.getManifest().version;
-  }
-  async tryGetTab(t) {
-    try {
-      return await chrome.tabs.get(Number.parseInt(t));
-    } catch {
-      return null;
-    }
-  }
-  getOsType() {
-    const t = navigator.userAgent;
-    const n = navigator.platform;
-    const a = ['Macintosh', 'MacIntel', 'MacPPC', 'Mac68K', 'macOS'];
-    const i = ['Win32', 'Win64', 'Windows', 'WinCE'];
-    const s = ['iPhone', 'iPad', 'iPod'];
-    return a.indexOf(n) !== -1
-      ? 'OSX'
-      : s.indexOf(n) !== -1
-        ? 'IOS'
-        : i.indexOf(n) !== -1
-          ? 'Windows'
-          : /Android/.test(t)
-            ? 'Android'
-            : /Linux/.test(n)
-              ? 'Linux'
-              : (() => {
-                  console.error('unable to detect os type, use Windows as default', n, t);
-                  return 'Windows';
-                })();
-  }
-  isChrome() {
-    const t = navigator.userAgent;
-    const n = t.includes('Chrome') && t.includes('Safari');
-    const a = t.includes('Edg');
-    const i = t.includes('OPR');
-    return n && !a && !i;
-  }
-
-  getDefaultShortcutKey() {
-    return 'b';
-  }
-
-  getDefaultSendShortcutKey() {
-    return 'enter';
-  }
-}
-
-export const reflyEnv = new ReflyEnv();
+export const serverOrigin =
+  import.meta.env.VITE_API_URL || (typeof window !== 'undefined' && window.ENV?.API_URL);
+export const wsServerOrigin =
+  import.meta.env.VITE_COLLAB_URL || (typeof window !== 'undefined' && window.ENV?.COLLAB_URL);
