+++ packages/utils/package.json
@@ -56,6 +56,7 @@
   "devDependencies": {
     "@types/chrome": "0.0.227",
     "@types/react": "^18.2.56",
-    "@types/react-dom": "^18.2.19"
+    "@types/react-dom": "^18.2.19",
+    "@refly/common-types": "workspace:*"
   }
 }
